ui: port various functionality (bug 1870696)
- port template helpers
- port forms
- prep revisions endpoint for porting
- move additional legacy endpoints to own module
- remove license headers from templates
