+++ src/lando/ui/jinja2/partials/footer.html
@@ -1,9 +1,3 @@
-{#
-    This Source Code Form is subject to the terms of the Mozilla Public
-    License, v. 2.0. If a copy of the MPL was not distributed with this
-    file, You can obtain one at https://mozilla.org/MPL/2.0/.
-#}
-
 <footer class="Footer footer">
   <div class="container">
     <div class="content has-text-centered">
