+++ src/lando/ui/jinja2/signout.html
@@ -1,9 +1,3 @@
-{#
-    This Source Code Form is subject to the terms of the Mozilla Public
-    License, v. 2.0. If a copy of the MPL was not distributed with this
-    file, You can obtain one at https://mozilla.org/MPL/2.0/.
-#}
-
 {% extends "partials/layout.html" %}
 {% block page_title %}Signed Out - Lando - Mozilla{% endblock %}
 {% block main %}
