--- src/lando/ui/jinja2/stack/partials/revision-author.html
@@ -1,7 +1,0 @@
-{#
-    This Source Code Form is subject to the terms of the Mozilla Public
-    License, v. 2.0. If a copy of the MPL was not distributed with this
-    file, You can obtain one at https://mozilla.org/MPL/2.0/.
-#}
-
-
