+++ src/lando/ui/jinja2/stack/partials/revision-reviewers.html
@@ -1,9 +1,3 @@
-{#
-    This Source Code Form is subject to the terms of the Mozilla Public
-    License, v. 2.0. If a copy of the MPL was not distributed with this
-    file, You can obtain one at https://mozilla.org/MPL/2.0/.
-#}
-
 <ul class="StackPage-reviewer-list">
   {%- for reviewer in reviewers %}
     <li class="{{ reviewer|reviewer_to_status_badge_class }} StackPage-reviewer-list-badge">
