+++ src/lando/ui/jinja2/stack/partials/timeline.html
@@ -1,9 +1,3 @@
-{#
-    This Source Code Form is subject to the terms of the Mozilla Public
-    License, v. 2.0. If a copy of the MPL was not distributed with this
-    file, You can obtain one at https://mozilla.org/MPL/2.0/.
-#}
-
 <div class="StackPage-timeline">
     {% if transplants %}
     {%- for transplant in transplants|sort(attribute='updated_at', reverse=True) %}
