+++ src/lando/ui/views.py
@@ -1,3 +1,5 @@
-from django.shortcuts import render
+from django.views import View
 
-# Create your views here.
+
+class LandoView(View):
+    pass
