+++ src/lando/urls.py
@@ -17,7 +17,10 @@ Including another URLconf
 from django.contrib import admin
 from django.urls import include, path
 
+from lando.ui.legacy import revisions
+
 urlpatterns = [
     path("admin/", admin.site.urls),
     path("oidc/", include("mozilla_django_oidc.urls")),
+    path("/D<int:revision_id>/", revisions.Revision.as_view()),
 ]
